refactor: 💡 move Timezones to keen.io/query
