+++ packages/ui-core/package.json
@@ -34,6 +34,7 @@
   "dependencies": {
     "@keen.io/colors": "^1.1.0",
     "@keen.io/icons": "^1.4.0",
+    "@keen.io/query": "^1.0.1",
     "color": "^3.1.2",
     "framer-motion": "^1.11.1",
     "fuse.js": "^6.4.1",
